+++ Sources/SecureStore/SecureStoreError.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public enum SecureStoreError: Error {
+    case unableToRetrieveFromUserDefaults
+    case cantGetPublicKeyFromPrivateKey
+    case cantStoreKey
+    case cantRetrieveKey
+    case cantEncryptData
+    case cantDecryptData
+}
